+++ src/layouts/InterfaceLayout/InterfaceLayout.vue
@@ -145,6 +145,7 @@ import {
   LedgerWallet,
   TrezorWallet,
   BitBoxWallet,
+  BitBox02Wallet,
   SecalotWallet,
   KeepkeyWallet
 } from '@/wallets';
@@ -153,6 +154,7 @@ import {
   LEDGER as LEDGER_TYPE,
   TREZOR as TREZOR_TYPE,
   BITBOX as BITBOX_TYPE,
+  BITBOX02 as BITBOX02_TYPE,
   SECALOT as SECALOT_TYPE,
   KEEPKEY as KEEPKEY_TYPE,
   MNEMONIC as MNEMONIC_TYPE
@@ -340,6 +342,26 @@ export default {
           break;
         case BITBOX_TYPE:
           this.togglePasswordModal(BitBoxWallet, 'BitBox');
+          break;
+        case BITBOX02_TYPE:
+          // eslint-disable-next-line no-case-declarations
+          let bb02;
+          BitBox02Wallet(this.$store.dispatch)
+            .then(_newWallet => {
+              bb02 = _newWallet;
+              this.$emit('bitbox02Open', bb02);
+              bb02
+                .init('')
+                .then(() => {
+                  this.toggleNetworkAddrModal(bb02);
+                })
+                .catch(e => {
+                  BitBox02Wallet.errorHandler(e);
+                });
+            })
+            .catch(e => {
+              BitBox02Wallet.errorHandler(e);
+            });
           break;
         case SECALOT_TYPE:
           this.togglePasswordModal(SecalotWallet, 'Secalot');