+++ src/wallets/hardware/index.js
@@ -1,10 +1,12 @@
 import BitBoxWallet from './bitbox';
+import BitBox02Wallet from './bitbox02';
 import LedgerWallet from './ledger';
 import TrezorWallet from './trezor';
 import SecalotWallet from './secalot';
 import KeepkeyWallet from './keepkey';
 export {
   BitBoxWallet,
+  BitBox02Wallet,
   LedgerWallet,
   TrezorWallet,
   SecalotWallet,