+++ src/wallets/index.js
@@ -2,6 +2,7 @@ import {
   LedgerWallet,
   TrezorWallet,
   BitBoxWallet,
+  BitBox02Wallet,
   SecalotWallet,
   KeepkeyWallet
 } from './hardware';
@@ -17,6 +18,7 @@ export {
   LedgerWallet,
   TrezorWallet,
   BitBoxWallet,
+  BitBox02Wallet,
   SecalotWallet,
   KeepkeyWallet,
   MewConnectWallet,