+++ tests/unit/specs/src/components/PopOver/PopOver.spec.js
@@ -45,6 +45,11 @@ describe('PopOver.vue', () => {
     });
   });
 
+  afterEach(() => {
+    wrapper.destroy();
+    wrapper = null;
+  });
+
   it('should render correct popcontent', () => {
     expect(
       wrapper.vm.$el.querySelector('p.popover-content').textContent