+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Release v5.4.4
+
 ### Feature
 
 - WalletConnect & WalletLink Support [#1924](https://github.com/MyEtherWallet/MyEtherWallet/pull/1924)
@@ -5,12 +7,15 @@
 
 ### Bug
 
+- Fix exit-to-fiat dropdown, iban indicator [#2134](https://github.com/MyEtherWallet/MyEtherWallet/pull/2134)
 - Fix button positions [#2125](https://github.com/MyEtherWallet/MyEtherWallet/pull/2125)
 
 ### Devop
 
 - Update BitBox Logo [#2120](https://github.com/MyEtherWallet/MyEtherWallet/pull/2120)
+- Update BitBox Logo [#2141](https://github.com/MyEtherWallet/MyEtherWallet/pull/2141)
 - Update greenkeeper packages [#2131](https://github.com/MyEtherWallet/MyEtherWallet/pull/2131)
+- Add security.txt to repo [#2140](https://github.com/MyEtherWallet/MyEtherWallet/pull/2140)
 
 ### Release v5.4.3
 
@@ -24,7 +29,7 @@
 
 ### Bugs
 
-- MEW Wallet modal ui changes [#2112] (https://github.com/MyEtherWallet/MyEtherWallet/pull/2112)
+- MEW Wallet modal ui changes [#2112](https://github.com/MyEtherWallet/MyEtherWallet/pull/2112)
 
 ### Release v5.4.2-hotfix.1
 
@@ -35,6 +40,7 @@
 ### Release v5.4.2
 
 ### Feature
+
 - MEW Wallet UI changes [#2048](https://github.com/MyEtherWallet/MyEtherWallet/pull/2048)
 
 ### Bugs